Add files via upload
One grid row has been built. This includes working s and o placement.
Done: working buttons for s and o placement and grid tiles. The grid as a whole will be straightforward to build by copy-pasting my code and making a few tweaks. This was actually fun to play with ^_^

Todo: Align everything perfectly, recolour buttons to look nice.
Matchmaking, player name and score, win-screen.
Almost there!

(after this is all over, the splash screen will be removed as it is not original code)